+++ semcore/data-table/__tests__/data-table-states.browser-test.tsx
@@ -682,6 +682,15 @@ test.describe('Limited state', () => {
       }
       await expect(locators.rowTableInTable(page, 2, 5).getByRole('gridcell').first()).not.toBeVisible();
     });
+
+    await test.step('Verify row doesn\'t expand accordion when it\'s under overlay', async () => {
+      const htmlContent = await e2eStandToHtml(standPath, 'en', { rowsLimit: 0, columnsLimit: 0 });
+      await page.setContent(htmlContent);
+
+      await locators.toggle(page).first().click({ force: true });
+      await locators.rowTableInTable(page, 2, 4).waitFor({ state: 'detached' });
+      await expect(locators.rowTableInTable(page, 2, 5).getByRole('gridcell').first()).not.toBeVisible();
+    });
   });
 });